+++ include/generator.h
@@ -11,18 +11,18 @@ class Generator {
     public:
         Segment generate_segment(
             int length,
-            float velocity = 2.0,
+            double velocity = 2.0,
             Coordinate start_c = {40.0, 10.0},
             int start_t = 0,
-            float start_e = 100.0);
+            double start_e = 100.0);
     
     private:
         Coordinate _start_coordinate;
         int _start_time;
-        float _start_elevation;
+        double _start_elevation;
 };
 
-Segment generate_track(std::vector<float> length_list, std::vector<float> velocity_list);
+Segment generate_track(std::vector<double> length_list, std::vector<double> velocity_list);
